Update escape/unescape in URL generation to support Ruby 3.x
In [a previous commit][1], we implemented a solution posted to the
now-read-only Paperclip repo. This failed the tests on the legacy app
where this fork is being used.

Instead, we're taking the implementation from [this commit][2] from
another fork of Paperclip which seems to work for both escape and
unescape.

[1]: https://github.com/maxshelley/paperclip/blob/925d24a4ee4420e3223b85ab6dff05fb02762a24/lib/paperclip/url_generator.rb
[2]: https://github.com/kreeti/kt-paperclip/commit/f0367733ddf37141ae00517c6697cc3dd233dab9